+++ src/Control/Effect/RWS/Strict.hs
@@ -34,7 +34,10 @@ import Data.Coerce   (coerce)
 import qualified Control.Monad.Trans.RWS.CPS as RWS
 
 import Control.Effect.Machinery
-import Control.Effect.RWS (RWS, RWS')
+import Control.Effect.Reader    (Reader, Reader')
+import Control.Effect.RWS       (RWS, RWS')
+import Control.Effect.State     (State, State')
+import Control.Effect.Writer    (Writer, Writer')
 
 -- This is necessary until the writer CPS instances land in monad-control.
 -- See: https://github.com/basvandijk/monad-control/pull/51
@@ -47,7 +50,7 @@ newtype RWST r w s m a =
   RWST { runRWST :: RWS.RWST r w s m a }
     deriving (Applicative, Functor, Monad, MonadIO)
     deriving (MonadTrans)
-    deriving (RWS' tag r w s)
+    deriving (RWS' tag r w s, Reader' tag r, Writer' tag w, State' tag s)
 
 instance MonadBase b m => MonadBase b (RWST r w s m) where
   liftBase = liftBaseDefault
@@ -70,52 +73,67 @@ instance Monoid w => MonadTransControl (RWST r w s) where
   {-# INLINABLE restoreT #-}
 
 -- | Runs the RWS effect and discards the final state.
-evalRWS' :: forall tag r w s m a. (Functor m, Monoid w)
-         => r                                     -- ^ The initial environment.
-         -> s                                     -- ^ The initial state.
-         -> (RWS' tag r w s `Via` RWST r w s) m a -- ^ The program whose RWS effect should be handled.
-         -> m (w, a)                              -- ^ The program with its RWS effect handled, producing the final
-                                                  -- output @w@ and the result @a@.
+evalRWS'
+  :: forall tag r w s m a. (Functor m, Monoid w)
+  => r
+  -- ^ The initial environment.
+  -> s
+  -- ^ The initial state.
+  -> ('[RWS' tag r w s, Reader' tag r, Writer' tag w, State' tag s] `EachVia` RWST r w s) m a
+  -- ^ The program whose RWS effect should be handled.
+  -> m (w, a)
+  -- ^ The program with its RWS effect handled, producing the final
+  -- output @w@ and the result @a@.
 evalRWS' r s = fmap reorder . (\m -> RWS.runRWST m r s) . coerce
   where
     reorder (a, _, w) = (w, a)
 {-# INLINE evalRWS' #-}
 
 -- | The untagged version of 'evalRWS''.
-evalRWS :: (Functor m, Monoid w) => r -> s -> (RWS r w s `Via` RWST r w s) m a -> m (w, a)
+evalRWS :: (Functor m, Monoid w) => r -> s -> ('[RWS r w s, Reader r, Writer w, State s] `EachVia` RWST r w s) m a -> m (w, a)
 evalRWS = evalRWS' @G
 {-# INLINE evalRWS #-}
 
 -- | Runs the RWS effect and discards the result of the interpreted program.
-execRWS' :: forall tag r w s m a. (Functor m, Monoid w)
-         => r                                     -- ^ The initial environment.
-         -> s                                     -- ^ The initial state.
-         -> (RWS' tag r w s `Via` RWST r w s) m a -- ^ The program whose RWS effect should be handled.
-         -> m (w, s)                              -- ^ The program with its RWS effect handled, producing the final
-                                                  -- output @w@ and the final state @s@.
+execRWS'
+  :: forall tag r w s m a. (Functor m, Monoid w)
+  => r
+  -- ^ The initial environment.
+  -> s
+  -- ^ The initial state.
+  -> ('[RWS' tag r w s, Reader' tag r, Writer' tag w, State' tag s] `EachVia` RWST r w s) m a
+  -- ^ The program whose RWS effect should be handled.
+  -> m (w, s)
+  -- ^ The program with its RWS effect handled, producing the final
+  -- output @w@ and the final state @s@.
 execRWS' r s = fmap reorder . (\m -> RWS.runRWST m r s) . coerce
   where
     reorder (_, s', w) = (w, s')
 {-# INLINE execRWS' #-}
 
 -- | The untagged version of 'execRWS''.
-execRWS :: (Functor m, Monoid w) => r -> s -> (RWS r w s `Via` RWST r w s) m a -> m (w, s)
+execRWS :: (Functor m, Monoid w) => r -> s -> ('[RWS r w s, Reader r, Writer w, State s] `EachVia` RWST r w s) m a -> m (w, s)
 execRWS = execRWS' @G
 {-# INLINE execRWS #-}
 
 -- | Runs the RWS effect and returns the final output, the final state and the result of the interpreted program.
-runRWS' :: forall tag r w s m a. (Functor m, Monoid w)
-        => r                                     -- ^ The initial environment.
-        -> s                                     -- ^ The initial state.
-        -> (RWS' tag r w s `Via` RWST r w s) m a -- ^ The program whose RWS effect should be handled.
-        -> m (w, s, a)                           -- ^ The program with its RWS effect handled, producing the final
-                                                 -- output @w@, the final state @s@ and the result @a@.
+runRWS'
+  :: forall tag r w s m a. (Functor m, Monoid w)
+  => r
+  -- ^ The initial environment.
+  -> s
+  -- ^ The initial state.
+  -> ('[RWS' tag r w s, Reader' tag r, Writer' tag w, State' tag s] `EachVia` RWST r w s) m a
+  -- ^ The program whose RWS effect should be handled.
+  -> m (w, s, a)
+  -- ^ The program with its RWS effect handled, producing the final
+  -- output @w@, the final state @s@ and the result @a@.
 runRWS' r s = fmap reorder . (\m -> RWS.runRWST m r s) . coerce
   where
     reorder (a, s', w) = (w, s', a)
 {-# INLINE runRWS' #-}
 
 -- | The untagged version of 'runRWS''.
-runRWS :: (Functor m, Monoid w) => r -> s -> (RWS r w s `Via` RWST r w s) m a -> m (w, s, a)
+runRWS :: (Functor m, Monoid w) => r -> s -> ('[RWS r w s, Reader r, Writer w, State s] `EachVia` RWST r w s) m a -> m (w, s, a)
 runRWS = runRWS' @G
 {-# INLINE runRWS #-}